+++ src/main.rs
@@ -1,6 +1,7 @@
 mod vm;
 mod write;
 mod lexer;
+mod parser;
 fn main() {
     write::println_to_console(b"Hello, World!");
 }
